refactor(p2): extract ChatListIndicators into sibling file (stage 2 of 4)
Moves the 5 indicator / marker composables (L254-609 of post-stage-1
ChatList.kt) into the new sibling ChatListIndicators.kt in the same
package: TimelineHistoryLoadingIndicator, AgentWorkingIndicator,
ContextCompactInProgressMarker, ContextCompactMarker, PendingUserMessageBubble.

Byte-equal: bodies and signatures unchanged.

Visibility widenings (5 sibling p→i, for ChatListNormal still in coord
to keep calling): TimelineHistoryLoadingIndicator, AgentWorkingIndicator,
ContextCompactInProgressMarker, ContextCompactMarker, PendingUserMessageBubble.

Visibility widenings (2 coord p→i, for sibling access to shared file-private
helpers): COMPACT_SUMMARY_WHITESPACE_RE (used by ContextCompactInProgressMarker's
proseOnly remember block), dashedRoundedBorder (used by PendingUserMessageBubble's
Surface modifier chain).

Coordinator strip: 2010 -> 1642 (-368 with import cleanup).
Imports trimmed: 11 orphans — Package01, LinearEasing, RepeatMode, animateFloat,
infiniteRepeatable, rememberInfiniteTransition, BoxWithConstraints, Brush,
PendingUserMessageMode, previewText, PigLoadingIndicator — all verified body
refs only in extracted range.

Pre-existing dead imports preserved verbatim.

Baseline 3/467 unit tests preserved.
